requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
nltk==3.8.1
pandas==2.0.3
matplotlib==3.7.2
textblob==0.17.1
newspaper3k==0.2.8
tqdm==4.65.0